further separate db and public models
