+++ backend/libs/core/src/core/doc_mgr/doc/delete.py
@@ -1,16 +1,12 @@
 import logging
-import uuid
 
-from core_db.db_models import DbTrackedDocument
-from core_db.providers.sql_database import DataDomain, get_sessionmaker
-from sqlalchemy import delete
+from core_db.doc_mgr.doc.delete import delete_tracking_record
+from core_db.doc_mgr.doc.query import get_documents
 
 from ...providers.file_store import get_s3_bucket
 from ...providers.vector_store import delete_vectors_by_document_id
-from .query import get_documents
 
 logger = logging.getLogger(__name__)
-
 
 def delete_document(document_id):
     """Delete tracking record, document from file store, and embeddings from vector store."""
@@ -40,23 +36,8 @@ def delete_document(document_id):
 
     # Delete tracking record.
 
-    _delete_tracking_record(tracking_record.id)
+    delete_tracking_record(tracking_record.id)
 
     return success
 
 
-def _delete_tracking_record(doc_uuid):
-    """Deletes the tracking record."""
-    if isinstance(doc_uuid, str):
-        doc_uuid = uuid.UUID(hex=doc_uuid)
-
-    sessionmaker = get_sessionmaker(DataDomain.ANSWERS)
-
-    with sessionmaker() as session, session.begin():
-        stmt = delete(DbTrackedDocument).where(DbTrackedDocument.id == doc_uuid)
-        result = session.execute(stmt)
-
-    if result.rowcount == 0:
-        logger.warning('No tracking record found with UUID %s', doc_uuid)
-    else:
-        logger.debug('Successfully deleted tracking record with UUID %s', doc_uuid)
