+++ backend/libs/core/src/core/doc_mgr/doc_set/__init__.py
@@ -1,6 +1,7 @@
+from core_db.doc_mgr.doc_set.query import get_document_sets
 from .add import add_document_set
 from .delete import delete_document_set
-from .query import get_document_sets, list_document_sets
+from .query import list_document_sets
 from .stats import get_document_set_statistics
 from .update import update_document_set
