+++ backend/libs/core/src/core/doc_mgr/doc_set/add.py
@@ -1,11 +1,8 @@
 import logging
 import uuid
 
-from core_db.db_models import DbTrackedDocumentSet
-from core_db.providers.sql_database import DataDomain, get_sessionmaker
-from sqlalchemy import select
+from core_db.doc_mgr.doc_set.add import add_or_update_document_set
 
-from ...lib_config import get_lib_config
 
 logger = logging.getLogger(__name__)
 
@@ -17,43 +14,8 @@ def add_document_set(name: str, is_new_doc_default: bool, is_public_viewable: bo
     Creates a new document set entry in the database with the provided name, default status, public
     visibility, and user ownership information.
     """
-    return _add_or_update_document_set(
+    return add_or_update_document_set(
         name=name, is_new_doc_default=is_new_doc_default, is_public_viewable=is_public_viewable, user_id=user_id
     )
 
 
-def _add_or_update_document_set(name: str, is_new_doc_default: bool, is_public_viewable: bool, user_id: uuid.UUID):
-    """Adds or updates the document set."""
-    sessionmaker = get_sessionmaker(DataDomain.ANSWERS)
-
-    with sessionmaker() as session:
-        with session.begin():
-            stmt = select(DbTrackedDocumentSet).where(DbTrackedDocumentSet.name == name)
-            result = session.execute(stmt)
-            existing_obj = result.scalar_one_or_none()
-
-        with session.begin():
-            doc_root_dir = get_lib_config().doc_root_dir
-            s3_rel_path = doc_root_dir + '/' + name
-
-            if existing_obj:
-                doc_set_uuid = existing_obj.id
-                existing_obj.name = name
-                existing_obj.s3_rel_path = s3_rel_path
-                existing_obj.is_new_doc_default = is_new_doc_default
-                existing_obj.is_public_viewable = is_public_viewable
-                existing_obj.last_user_id = user_id
-            else:
-                doc_set_uuid = uuid.uuid4()
-
-                new_obj = DbTrackedDocumentSet(
-                    id=doc_set_uuid,
-                    name=name,
-                    s3_rel_path=s3_rel_path,
-                    is_new_doc_default=is_new_doc_default,
-                    is_public_viewable=is_public_viewable,
-                    last_user_id=user_id,
-                )
-                session.add(new_obj)
-
-    return doc_set_uuid
