+++ backend/libs/core/src/core/prompt_mgr/prompt/add.py
@@ -1,10 +1,8 @@
 import logging
 import uuid
 
-from core_db.db_models import DbAgentPrompt
-from core_db.providers.sql_database import DataDomain, get_sessionmaker
+from core_db.prompt_mgr.prompt.add import add_or_update_agent_prompt
 from core_public import AgentPromptStatus, OwnerType
-from sqlalchemy import and_, select, update
 
 logger = logging.getLogger(__name__)
 
@@ -24,7 +22,7 @@ def add_prompt(
     Creates a new prompt entry with automatic version incrementing. If the status is ACTIVE,
     deactivates all other versions of the same prompt name.
     """
-    _add_or_update_agent_prompt(
+    add_or_update_agent_prompt(
         name=name,
         owner_type=owner_type,
         status=status,
@@ -35,57 +33,3 @@ def add_prompt(
     )
 
 
-def _add_or_update_agent_prompt(
-    name: str,
-    owner_type: OwnerType,
-    status: AgentPromptStatus,
-    system_message: str | None,
-    human_message: str | None,
-    include_history: bool | None,
-    user_id: uuid.UUID,
-):
-    """Adds or updates the prompt."""
-    sessionmaker = get_sessionmaker(DataDomain.ANSWERS)
-
-    with sessionmaker() as session:
-        # We want to know auto-increment the version number.
-        # So this query will return the highest version numbered first.
-        with session.begin():
-            stmt = (
-                select(DbAgentPrompt).where(DbAgentPrompt.name == name).limit(1).order_by(DbAgentPrompt.version.desc())
-            )
-            result = session.execute(stmt)
-            existing_obj = result.scalar_one_or_none()
-
-            # IMPORTANT!!
-            # We should always access SQLAlchemy object properties inside a transaction. Its ORM
-            # has subtle lazy-loading behaviors, including when expire_on_commit=True (which is
-            # important for data consistency checking). Doing this will prevent auto-transactions
-            # from interferring with the next transaction.
-            version = 1 if existing_obj is None else (existing_obj.version + 1)
-
-        with session.begin():
-            prompt_uuid = uuid.uuid4()
-
-            new_obj = DbAgentPrompt(
-                id=prompt_uuid,
-                name=name,
-                owner_type=owner_type,
-                status=status,
-                system_message=system_message,
-                human_message=human_message,
-                include_history=include_history,
-                version=version,
-                last_user_id=user_id,
-            )
-            session.add(new_obj)
-
-        # Only one version can be active
-        if status == AgentPromptStatus.ACTIVE:
-            with session.begin():
-                stmt = (
-                    update(DbAgentPrompt)
-                    .where(and_(DbAgentPrompt.name == name, DbAgentPrompt.version != version))
-                    .values(status=AgentPromptStatus.DEACTIVATED)
-                )
-                result = session.execute(stmt)
