+++ backend/libs/core_db/src/core_db/doc_mgr/doc_set/add.py
@@ -1,0 +1,44 @@
+from core.lib_config import get_lib_config
+from core_db.db_models import DbTrackedDocumentSet
+from core_db.providers.sql_database import DataDomain, get_sessionmaker
+from sqlalchemy import select
+
+
+import uuid
+
+
+def add_or_update_document_set(name: str, is_new_doc_default: bool, is_public_viewable: bool, user_id: uuid.UUID):
+    """Adds or updates the document set."""
+    sessionmaker = get_sessionmaker(DataDomain.ANSWERS)
+
+    with sessionmaker() as session:
+        with session.begin():
+            stmt = select(DbTrackedDocumentSet).where(DbTrackedDocumentSet.name == name)
+            result = session.execute(stmt)
+            existing_obj = result.scalar_one_or_none()
+
+        with session.begin():
+            doc_root_dir = get_lib_config().doc_root_dir
+            s3_rel_path = doc_root_dir + '/' + name
+
+            if existing_obj:
+                doc_set_uuid = existing_obj.id
+                existing_obj.name = name
+                existing_obj.s3_rel_path = s3_rel_path
+                existing_obj.is_new_doc_default = is_new_doc_default
+                existing_obj.is_public_viewable = is_public_viewable
+                existing_obj.last_user_id = user_id
+            else:
+                doc_set_uuid = uuid.uuid4()
+
+                new_obj = DbTrackedDocumentSet(
+                    id=doc_set_uuid,
+                    name=name,
+                    s3_rel_path=s3_rel_path,
+                    is_new_doc_default=is_new_doc_default,
+                    is_public_viewable=is_public_viewable,
+                    last_user_id=user_id,
+                )
+                session.add(new_obj)
+
+    return doc_set_uuid
