+++ backend/libs/core_db/src/core_db/prompt_mgr/prompt/stats.py
@@ -1,0 +1,13 @@
+from core_db.db_models import DbAgentPrompt
+from core_db.providers.sql_database import DataDomain, get_sessionmaker
+from sqlalchemy import func, select
+
+
+def get_agent_prompt_stats():
+    """Return the count of records and maximum updated_date time in the prompt table."""
+    sessionmaker = get_sessionmaker(DataDomain.ANSWERS)
+
+    with sessionmaker() as session:
+        stmt = select(func.count().label('prompt_count'), func.max(DbAgentPrompt.update_time).label('max_update_time'))
+        result = session.execute(stmt).first()
+    return {'prompt_count': result[0], 'max_update_time': result[1]}
